+++ tests/src/test/scala/runtime/actionContainers/NodeJsActionContainerTests.scala
@@ -585,4 +585,5 @@ abstract class NodeJsActionContainerTests extends BasicActionRunnerTests with Ws
         e shouldBe empty
     })
   }
+
 }